Problem: (CRO-247) Broadcast transaction endpoint is vulnerable to DoS attack (#264)
Solution: Added maximum size of binary encoded data for a transaction. Currently, the default value is 1 MB (same as tendermint's default for a transaction) which can be changed once we have a better understanding of different usecases we want to support and the size required for those.

Note: This change is temporary and should be reverted once tendermint officially starts supporting `max_tx_bytes` in `cargo.toml`.