Problem: (CRO-444) auto sync doesn't work correctly after receceiving incorrect messages
Solutoin:
update send queue after closing the session, which is the channel between core and network.
this is Arc::Mutex related bug, because it's shallow cloning,
we need to use &mut to change it
without that, mutex value is Option, it is just replaced on the stack
and destroyed after it's out of scope.