Problem: app hash changes even if app state didn't change (CRO-526)
Solution: added a flag to chainabci node to signal whether rewards pool was updated
in that block (either by slashing events or transaction fees -- in the future reward distribution),
so that its "last_block_height" (which denotes when it was updated last time)
is only updated when the amount changed.
TODO: perhaps move the rewards pool update logic to rewards pool code (rather than in abci)?