Problem: `dev-utils` cannot generate `genesis.json` for testnet
Solution: Added a new option in genesis generation command
(`unbonded_address`). If `unbonded_address` is provided, all other
staking addresses will be marked as `Bonded` except for the one provided
in `unbonded_address`, which will be marked as `UnbondedFromGenesis`.
If `unbonded_address` is not provided, everything will work the same.