Problem: no domain separation of hashed data types (fixes #1715)
Solution: currently the data types are distinct, but it could happen
in the future that two TX types have the same/similar schema,
so the txid would be the same and one may potentially "swap"
tx type (e.g. one may sign unjailtx, but malicious party
would broadcast it as unbondtx -- currently not possible,
bu just an example of what this aims to prevent).
- a trait "TransactionType" was added and implemented for all tx payloads
- txid is now calculated as blake3(<type tag>||scale_payload(tx))
- note this "breaks" 0.5 signing code
(as even though the payload is the same, wallets will be
using a different message digest) -- for that reason,
it's feature-guarded under "new-txid"