Modify verification code
Add a bunch of `trace!` statements and refactor the verification logic.
Verification happens on both client and server side, we control the
verification logic with `mr_enclave`, and `previous_mr_enclave`.

- With no `mr_enclave` we verify if:
  - `isv_svn`s match

- With `mr_enclave` we verify if:
  - `isv_svn`s match and `mr_enclave`s match
  - peer's `isv_svn` is one bigger than ours (no check on `mr_enclave`,
    this is used during upgrade by the server).

- With `previous_mr_enclave` we verify if:
  - Our `isv_svn` is one bigger than peer and the `previous_mr_enclave`
    is equal to the `mr_enclave` of peer (this is used by client during
    upgrade).

Said another way, this PR allows an upgraded instance to connect to an
instance running the previous svn versioned software.